[av1d] Switch from Rev0.25.2 to Rev0.5
Issue: VSMGWL-10945
Test: 134 of 135 Rev0.5 test streams are decoded b2b with C-model

One failing test case is confirmed fulsim issue.

+++ _studio/shared/umc/codec/av1_dec/src/umc_av1_bitstream.cpp
@@ -1523,7 +1523,7 @@ namespace UMC_AV1_DECODER
 
     void AV1Bitstream::ReadTile(FrameHeader const* fh, size_t& reportedSize, size_t& actualSize)
     {
-        actualSize = reportedSize = GetLE(fh->tileSizeBytes);
+        actualSize = reportedSize = static_cast<size_t>(GetLE(fh->tileSizeBytes));
 
         if (BytesLeft() < reportedSize)
             actualSize = BytesLeft();
